# report/openapi.json
{"openapi":"3.1.0","info":{"title":"财报收集API服务","description":"\n中国A股、港股财报PDF收集与提取API\n\n## 功能\n\n* **搜索**: 搜索A股和港股上市公司财报\n* **下载**: 下载单个或批量财报PDF\n* **查询**: 查询已下载PDF信息和统计\n* **提取**: 提取PDF中的财务数据、表格和文本\n* **缓存**: 管理提取结果缓存\n\n## 数据源\n\n* A股: 巨潮资讯网 (cninfo.com.cn)\n* 港股: 港交所披露易 (hkexnews.hk)\n        ","version":"1.0.0"},"paths":{"/api/v1/reports/cn/search":{"get":{"tags":["搜索"],"summary":"Search Cn Reports","description":"搜索中国A股上市公司财报\n\n- **stock_code**: 6位数字股票代码 (如: 000001, 600519)\n- **report_type**: 报告类型\n- **max_count**: 最大返回数量 (1-50)\n\n数据源: 巨潮资讯网","operationId":"search_cn_reports_api_v1_reports_cn_search_get","parameters":[{"name":"stock_code","in":"query","required":true,"schema":{"type":"string","pattern":"^[0-9]{6}$","description":"6位股票代码","title":"Stock Code"},"description":"6位股票代码"},{"name":"report_type","in":"query","required":false,"schema":{"type":"string","description":"报告类型: annual(年报), semi_annual(半年报), quarterly(季报), all(全部)","default":"annual","title":"Report Type"},"description":"报告类型: annual(年报), semi_annual(半年报), quarterly(季报), all(全部)"},{"name":"max_count","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"最大返回数量","default":10,"title":"Max Count"},"description":"最大返回数量"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/hk/search":{"get":{"tags":["搜索"],"summary":"Search Hk Reports","description":"搜索港股上市公司财报\n\n- **stock_code**: 5位数字股票代码 (如: 00700, 01810)\n- **report_type**: 报告类型\n- **max_count**: 最大返回数量 (1-50)\n\n数据源: 港交所披露易","operationId":"search_hk_reports_api_v1_reports_hk_search_get","parameters":[{"name":"stock_code","in":"query","required":true,"schema":{"type":"string","pattern":"^[0-9]{5}$","description":"5位股票代码","title":"Stock Code"},"description":"5位股票代码"},{"name":"report_type","in":"query","required":false,"schema":{"type":"string","description":"报告类型: annual(年报), semi_annual(中期报告), quarterly(季报)","default":"annual","title":"Report Type"},"description":"报告类型: annual(年报), semi_annual(中期报告), quarterly(季报)"},{"name":"max_count","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"最大返回数量","default":10,"title":"Max Count"},"description":"最大返回数量"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/cn/download":{"post":{"tags":["下载"],"summary":"Download Cn Report","description":"下载单个A股财报PDF\n\n- **stock_code**: 6位股票代码\n- **url**: 从搜索API获取的下载URL\n- **title**: 可选的文件标题","operationId":"download_cn_report_api_v1_reports_cn_download_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_download_cn_report_api_v1_reports_cn_download_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/cn/batch-download":{"post":{"tags":["下载"],"summary":"Batch Download Cn Reports","description":"批量下载A股财报PDF\n\n自动搜索并下载指定股票的财报","operationId":"batch_download_cn_reports_api_v1_reports_cn_batch_download_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDownloadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/hk/batch-download":{"post":{"tags":["下载"],"summary":"Batch Download Hk Reports","description":"批量下载港股财报PDF\n\n自动搜索并下载指定股票的财报","operationId":"batch_download_hk_reports_api_v1_reports_hk_batch_download_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDownloadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pdfs":{"get":{"tags":["下载"],"summary":"List Downloaded Pdfs","description":"列出已下载的PDF文件\n\n支持按股票代码、市场、报告类型筛选","operationId":"list_downloaded_pdfs_api_v1_pdfs_get","parameters":[{"name":"stock_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"股票代码筛选","title":"Stock Code"},"description":"股票代码筛选"},{"name":"market","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"市场筛选 (CN/HK/US)","title":"Market"},"description":"市场筛选 (CN/HK/US)"},{"name":"report_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"报告类型筛选","title":"Report Type"},"description":"报告类型筛选"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"返回数量","default":20,"title":"Limit"},"description":"返回数量"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pdfs/{pdf_id}":{"get":{"tags":["查询"],"summary":"Get Pdf Info","description":"获取单个PDF详细信息\n\n- **pdf_id**: PDF记录ID (从列表API获取)","operationId":"get_pdf_info_api_v1_pdfs__pdf_id__get","parameters":[{"name":"pdf_id","in":"path","required":true,"schema":{"type":"integer","description":"PDF记录ID","title":"Pdf Id"},"description":"PDF记录ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stats":{"get":{"tags":["查询"],"summary":"Get Collection Stats","description":"获取收集统计信息\n\n返回各市场PDF数量、总文件大小等统计","operationId":"get_collection_stats_api_v1_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}}}}},"/api/v1/pdfs/cleanup":{"delete":{"tags":["查询"],"summary":"Cleanup Old Pdfs","description":"清理旧PDF文件\n\n- **days**: 清理多少天前下载的文件","operationId":"cleanup_old_pdfs_api_v1_pdfs_cleanup_delete","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"description":"清理多少天前的文件","default":90,"title":"Days"},"description":"清理多少天前的文件"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/extract/content":{"post":{"tags":["提取"],"summary":"Extract Pdf Content","description":"提取PDF中的结构化财务数据\n\n提供 pdf_path 或 pdf_id 之一:\n- **pdf_path**: PDF文件绝对路径\n- **pdf_id**: 从列表API获取的PDF记录ID\n- **force_refresh**: 是否强制重新提取(忽略缓存)\n\n自动识别并提取:\n- 利润表 (营业收入、净利润等)\n- 资产负债表 (总资产、负债、权益等)\n- 现金流量表 (经营、投资、筹资现金流等)\n- 财务指标 (EPS、ROE、负债率等)\n\n支持缓存，重复提取将直接返回缓存结果","operationId":"extract_pdf_content_api_v1_extract_content_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_extract_pdf_content_api_v1_extract_content_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/extract/tables":{"post":{"tags":["提取"],"summary":"Extract Pdf Tables","description":"提取PDF中的表格数据\n\n提供 pdf_path 或 pdf_id 之一\n\n返回PDF中识别出的所有表格，以列表形式呈现","operationId":"extract_pdf_tables_api_v1_extract_tables_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_extract_pdf_tables_api_v1_extract_tables_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/extract/text":{"post":{"tags":["提取"],"summary":"Extract Pdf Text","description":"提取PDF全部文本\n\n提供 pdf_path 或 pdf_id 之一\n\n返回PDF中的纯文本内容","operationId":"extract_pdf_text_api_v1_extract_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_extract_pdf_text_api_v1_extract_text_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cache/stats":{"get":{"tags":["缓存"],"summary":"Get Cache Stats","description":"获取缓存统计信息\n\n返回缓存条目数量、大小、版本等信息","operationId":"get_cache_stats_api_v1_cache_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}}}}},"/api/v1/cache/cleanup":{"delete":{"tags":["缓存"],"summary":"Cleanup Cache","description":"清理过期缓存\n\n删除指定天数之前的提取结果缓存","operationId":"cleanup_cache_api_v1_cache_cleanup_delete","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"description":"清理多少天前的缓存","default":90,"title":"Days"},"description":"清理多少天前的缓存"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cache/warm":{"post":{"tags":["缓存"],"summary":"Warm Cache","description":"预热缓存\n\n对未缓存的PDF提前进行内容提取并缓存结果\n\n- **stock_code**: 可选，限定股票代码\n- **market**: 可选，限定市场\n- **max_files**: 最大处理文件数","operationId":"warm_cache_api_v1_cache_warm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WarmCacheRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Dict_str__Any__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["系统"],"summary":"Health Check","description":"健康检查","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["系统"],"summary":"Root","description":"根路径","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"APIResponse_Dict_str__Any__":{"properties":{"success":{"type":"boolean","title":"Success","description":"操作是否成功"},"data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Data","description":"返回数据"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error","description":"错误信息"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message","description":"提示信息"}},"type":"object","required":["success"],"title":"APIResponse[Dict[str, Any]]"},"BatchDownloadRequest":{"properties":{"stock_code":{"type":"string","title":"Stock Code","description":"股票代码"},"report_type":{"type":"string","pattern":"^(annual|semi_annual|quarterly|all)$","title":"Report Type","description":"报告类型","default":"annual"},"years":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Years","description":"年份列表"},"max_count":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Max Count","description":"最大下载数量","default":5}},"type":"object","required":["stock_code"],"title":"BatchDownloadRequest","description":"批量下载请求"},"Body_download_cn_report_api_v1_reports_cn_download_post":{"properties":{"stock_code":{"type":"string","title":"Stock Code","description":"股票代码"},"url":{"type":"string","title":"Url","description":"下载URL"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"文件标题"}},"type":"object","required":["stock_code","url"],"title":"Body_download_cn_report_api_v1_reports_cn_download_post"},"Body_extract_pdf_content_api_v1_extract_content_post":{"properties":{"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path","description":"PDF文件路径"},"pdf_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pdf Id","description":"PDF记录ID"},"force_refresh":{"type":"boolean","title":"Force Refresh","description":"强制重新提取(忽略缓存)","default":false}},"type":"object","title":"Body_extract_pdf_content_api_v1_extract_content_post"},"Body_extract_pdf_tables_api_v1_extract_tables_post":{"properties":{"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path","description":"PDF文件路径"},"pdf_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pdf Id","description":"PDF记录ID"}},"type":"object","title":"Body_extract_pdf_tables_api_v1_extract_tables_post"},"Body_extract_pdf_text_api_v1_extract_text_post":{"properties":{"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path","description":"PDF文件路径"},"pdf_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pdf Id","description":"PDF记录ID"}},"type":"object","title":"Body_extract_pdf_text_api_v1_extract_text_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WarmCacheRequest":{"properties":{"stock_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stock Code","description":"股票代码"},"market":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market","description":"市场"},"max_files":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"Max Files","description":"最大文件数","default":10}},"type":"object","title":"WarmCacheRequest","description":"预热缓存请求"}}}}